[BP] Map servers / sources / metadata template identifiers - delete confirmation dialog and fixes in display delete button in sources page. Related to #6641 (#6647)

+++ web-ui/src/main/resources/catalog/js/admin/SourcesController.js
@@ -161,9 +161,11 @@
                 });
       };
 
+      $scope.deleteSourceConfig = function () {
+        $("#gn-confirm-remove-source").modal("show");
+      };
 
-
-      $scope.removeSource = function() {
+      $scope.confirmDeleteSourceConfig = function () {
         $http.delete('../api/sources/' + $scope.source.uuid)
             .success(function(data) {
               $rootScope.$broadcast('StatusUpdated', {
@@ -172,6 +174,7 @@
                 type: 'success'});
 
               loadSources();
+              $scope.source = null;
             })
             .error(function(data) {
                   $rootScope.$broadcast('StatusUpdated', {
